role_relations

+++ database/migrations/2017_12_25_142322_create_articles_table.php
@@ -29,7 +29,7 @@ class CreateArticlesTable extends Migration
             $table->integer('created_by')->nullable();
             $table->integer('modified_by')->nullable();
             $table->timestamps();
-            $table->tinyInteger('user_id')->unsigned();
+            $table->integer('user_id')->unsigned();
 
             $table->foreign('user_id')
                 ->references('id')
